requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
aiofiles
aiogram
aiosqlite
